requirements.txt
openai == 1.3.7
pypdf == 3.17.4
python-dotenv == 1.0.0
langchain == 0.0.345
chromadb == 0.4.18
tiktoken == 0.5.2
django
djangorestframework == 3.14.0
tqdm == 4.66.1
django-cors-headers == 4.3.1
sentence-transformers == 2.2.2